move fn to Fn
