+++ src/aioway/fn/tensors.py
@@ -1,7 +1,6 @@
 # Copyright (c) AIoWay Authors - All Rights Reserved
 
 import abc
-import functools
 import operator
 import typing
 from collections import abc as cabc
@@ -89,73 +88,13 @@ class TensorFn(fn.Fn[torch.Tensor], abc.ABC):
         return UFunc2Thunk(operator.le, self, other)
 
     @typing.override
-    @typing.final
-    def do(self) -> torch.Tensor:
-        """
-        This is the public function that parent `Fn`s should call,
-        when the want to request the values of an `Fn`.
-
-        It handles caching in the normal case, so repeated calling the function means that
-        the expensive computation (defined within `forward`) would only be called once.
-
-        When fake mode is enabled, it calls `preview` for a fake tensor,
-        which is a preview for the normal computation to save computation cost.
-
-        The reason this is modal with `fake.is_enabled()` as a toggle,
-        to make sure `preview` and `forward` can use the same codepath as much as possible,
-        in the default case `preview` is `forward` with fake mode on.
-        """
-
-        if fake.is_enabled():
-            return self.preview()
-
-        else:
-            return self.__forward_cache()
-
-    @fake.enable_func
-    def preview(self) -> tsc.FakeTensor:
-        """
-        The `preview` function generates a "preview" for the `Tensor` that would be generated.
-
-        The result type (`FakeTensor`) is used as a worst case analysis of the original `Tensor`.
-
-        In most cases (non leaf operators), this method is just a clone of `forward`,
-        which is the default implementation of this function.
-
-        In the following cases it must be modified:
-
-        1. Source tensors, `forward` won't be `FakeTensor`, so conversion is needed.
-        2. Operators that cannot be supported by `torch` e.g. boolean  masking.
-        """
-
-        result = self.forward()
-        assert fake.is_fake_tensor(result)
-        return result
-
-    @abc.abstractmethod
-    def forward(self) -> torch.Tensor:
-        raise NotImplementedError
-
-    @typing.override
     @abc.abstractmethod
     def deps(self) -> tuple[fn.Fn[object], ...]:
         raise NotImplementedError
 
     @property
     def attr(self):
-        return schemas.attr(self.__get_fake_or_computed())
-
-    def __get_fake_or_computed(self):
-        "If `forward` has already been called, return it. Else return `preview`."
-
-        if self.__forward_cache.is_hit:
-            return self.__forward_cache()
-        else:
-            return self.preview()
-
-    @functools.cached_property
-    def __forward_cache(self):
-        return _TensorFnCache(self.forward)
+        return schemas.attr(self.forward() if self.done else self.preview())
 
     @property
     def shape(self):
@@ -183,38 +122,6 @@ def tensor(data: TensorFn | torch.Tensor) -> TensorFn:
         return TensorDataFn(data)
 
     raise TypeError(f"Do not know how to handle {type(data)=}.")
-
-
-@typing.final
-class _TensorFnCache:
-    """
-    The cacher for `TensorFn.forward`.
-
-    The reason we use this boilerplate over directly using `functools.cache`,
-    `functools.cached_property`, or having a saved `.__result` member for instance,
-    is because this is the least assuming.
-
-    `functools.cache` assumes that `self` is hashable.
-    `functools.cached_property` cannot inspect whether we have evaluated it or not.
-    `.__result` member assumes subclass calls `__init__` properly.
-
-    Since this is saved in a `functools.cached_property`, it can be used on unhashable types,
-    yet support inspecting whether we called it or not, and does not need to call `__init__`.
-    """
-
-    def __init__(self, func: cabc.Callable[[], torch.Tensor]) -> None:
-        self._result: torch.Tensor | None = None
-        self._func = func
-
-    def __call__(self) -> torch.Tensor:
-        if self._result is None:
-            self._result = self._func()
-
-        return self._result
-
-    @property
-    def is_hit(self):
-        return self._result is not None
 
 
 @_common.dcls_no_eq
